refactor(Request): Enhance `getRemoteIP()` method to validate the remote IP address and improve handling of server parameters. Add tests for retrieving remote IP from PSR-7 server parameters in `ServerParamsPsr7Test` class.
